Not to use syntax cache, only use func cache
